+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Refactoring `Classes\Tokenizer{}`.
 - Making it possible to tokenize specific shortcode tag names in `Classes\Core\Tokenizer{}`.
 - Adding `$compress` parameter to `c::normalizeEols()` to allow for compressino to be disabled there.
+- Adding `c::removeKey()` and `c::removeKeys()`.
 
 ## v160719.52827